+++ packages/components/molecules/f-rating/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.7.0
+------------------------------
+*November 01, 2022*
+
+### Added
+- Visual tests to cover changes in v0.6.0
+- `hasRatingAvailable` computed value to handle when to display visually hidden text.
+
+
 v0.6.0
 ------------------------------
 *October 26, 2022*
@@ -14,7 +23,7 @@ v0.6.0
 
 ### Changed
 - Split out from main rating file into Five & Single star components.
--
+
 
 v0.5.0
 ------------------------------